backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-genai==1.5.0
pymupdf==1.25.3
reportlab==4.2.5
python-dotenv==1.0.1
